Refactor: initializeをコンポーネントに切り出し
