+++ src/app/(play)/argutia/components/Argutia/argutia.module.css
@@ -23,29 +23,3 @@
   right: 0;
   z-index: 1000;
 }
-
-@keyframes fade {
-  0% {
-    opacity: 0;
-  }
-  20% {
-    opacity: 1;
-  }
-  80% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-  }
-}
-
-
-.fadeText {
-  color: var(--mantine-color-white);
-  font-size: 1.5rem;
-  font-weight: 700;
-  text-align: center;
-  margin-top: var(--mantine-spacing-xl);
-  margin-bottom: var(--mantine-spacing-xl);
-  animation: fade 5s infinite;
-}
