+++ src/app/(play)/argutia/components/Initialize/initialize.module.css
@@ -1,0 +1,36 @@
+.root {
+  background-color: #11284b;
+  background-size: cover;
+  background-position: center;
+  padding-top: var(--mantine-spacing-xl);
+  height: 100vh;
+  width: 100vw;
+  * {
+    color: var(--mantine-color-white);
+  }
+}
+
+@keyframes fade {
+  0% {
+    opacity: 0;
+  }
+  20% {
+    opacity: 1;
+  }
+  80% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0;
+  }
+}
+
+.fadeText {
+  color: var(--mantine-color-white);
+  font-size: 1.5rem;
+  font-weight: 700;
+  text-align: center;
+  margin-top: var(--mantine-spacing-xl);
+  margin-bottom: var(--mantine-spacing-xl);
+  animation: fade 5s infinite;
+}
